+++ lib/app/modules/news_detail/bindings/news_detail_binding.dart
@@ -1,0 +1,11 @@
+import 'package:get/get.dart';
+import '../controllers/news_detail_controller.dart';
+
+class NewsDetailBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<NewsDetailController>(
+      () => NewsDetailController(),
+    );
+  }
+}