+++ lib/app/routes/app_routes.dart
@@ -4,6 +4,7 @@ class Routes {
   static const DYNAMIC_FORM = '/dynamic-form';
   static const REPORT_HISTORY = '/report-history';
   static const REPORT_DETAIL = '/report-detail';
+  static const NEWS_DETAIL = '/news-detail';
 
   // Initial route
   static const INITIAL = LOGIN;